PROJ - Added shorts.lbr.
Unforunately, EAGLE does not allow for one netlist with multiple names. This is consistent with bringing in two signals together on a PCB, and directly connecting them. There are two solutions for this:

- Add additional silkscreen and text information to schematic/PCB respectively to show that although there is one common net name for two connectors, they should breakout to different parts of the circuit
- Use a sort of 'hack' by creating a part which directly shorts on a PCB. This second option produces a more obvious netlist and tidier schematic, but will set off DRC rules checks that must be manually approved.

I'll try out the former method to begin with.

The biggest problem is in using IC's with preset, implicit power rails (VCC, VEE, GND, VDD, VSS, etc.) In this case, we might want one power rail/net to satisfy multiple names. Here, we must either explicitly  every single power rail for each IC we use, or, we can go with the later option.

Simply put, there is no good option, but this library will be good for quick hotfixes in certain cases.