PROJ - Updated PF1262.lbr
Previous PF1262.lbr for precharge/discharge resistors was not compatible with ordered Riedon parts.
Unlike all other TO-126 packages, Riedon's resistors are 2 legged. Upon further inspection, this stated footprint is roughly compatible with the standard TO-126 package, with the middle terminal leg removed.

Two versions of this package were created. One with, and one with/out heatsink.