PROJ - Renamed library folders. Added con-MATE-N-LOK.
Custom library files were not showing up upon refresh in EAGLE. The best guess for why this behavior was occuring was that EAGLE only goes one folder deep in looking for matching library names. So, all custom libraries have been moved to a single parent folder.

Additionally, a con-MATE-N-LOK library has been added and will be added on as parts are needed. Please check for commits made on master branch for these.