PROJ - Added SC-STAGE.
We'll need to add in extra power-staging for all of our latching faults.
This can be easily achieved through MOSFETs. So, instead of using logic to filter out our control signal in combinational logic, we also use these individual latching faults to break the coil current before it reaches the AIR coils.